+++ R/tm_t_events_patyear.R
@@ -189,6 +189,7 @@ tm_t_events_patyear <- function(label,
                                 pre_output = NULL,
                                 post_output = NULL
                                 ) {
+  logger::log_info("Initializing tm_t_events_patyear")
   stop_if_not(
     is_character_single(dataname),
     is_character_single(parentname),