+++ R/tm_t_mult_events.R
@@ -320,6 +320,7 @@ tm_t_mult_events <- function(label, # nolint
                              drop_arm_levels = TRUE,
                              pre_output = NULL,
                              post_output = NULL) {
+  logger::log_info("Initializing tm_t_mult_events")
   stop_if_not(
     is_character_single(label),
     is_character_single(dataname),